+++ mobilesdk/src/main/java/com/swedbankpay/mobilesdk/paymentsession/PaymentSessionState.kt
@@ -1,7 +1,7 @@
 package com.swedbankpay.mobilesdk.paymentsession
 
-import android.view.View
 import androidx.annotation.Keep
+import androidx.fragment.app.Fragment
 import com.swedbankpay.mobilesdk.paymentsession.api.model.response.ProblemDetails
 import com.swedbankpay.mobilesdk.paymentsession.exposedmodel.AvailableInstrument
 import com.swedbankpay.mobilesdk.paymentsession.exposedmodel.PaymentSessionProblem
@@ -24,7 +24,7 @@ sealed class PaymentSessionState {
     /**
      * Called when 3d-secure should be shown
      */
-    class Show3dSecure(val view: View) : PaymentSessionState()
+    class Show3dSecure(val fragment: Fragment) : PaymentSessionState()
 
     /**
      * Called when 3d-secure can be dismissed
